fix GetMealRecommendation communication with FCE

+++ NutritionService/Features/GetMealRecommendations/Dtos/MealRecommendationResponseDto.cs
@@ -4,6 +4,6 @@ namespace NutritionService.Features.GetMealRecommendations.Dtos;
 
 public class MealRecommendationResponseDto
 {
-    public int UserDailyGoalCalories { get; init; }
+    public double UserDailyGoalCalories { get; init; }
     public PagedResponse<RecommendedMealDto>  RecommendedMeals { get; init; }
 }
